Removed the unnecessary generic typing from Point -- object values are now stored as primitive doubles.
This change results in minor differences in the Main class, where the doubles are occasionally casted to ints, and the diamond brackets are removed.

Two TODOs are added to Main.

The first involves creating a colorPostProcessor that will scale as MAX_ITERATIONS increases; the current postprocessor leaves colors washed out.

The second involves replacing the current coordinateMapper with a more generic coordinateMapper, which will eventually be provided in the Rectangle class